cleaning up, make more readable

+++ bin/scoutloader.pl
@@ -1,5 +1,5 @@
 #!/usr/bin/perl -w
-
+use Backticks;
 use strict;
 my $directory = '/fs1/results/cron/scout';
 
@@ -8,8 +8,8 @@ opendir (DIR, $directory) or die $!;
 while (my $file = readdir(DIR)) {
 
     if ( $file =~ /\.yaml/) {
-        my $fullpath = $directory."/".$file;
-        scoutcommand($fullpath);
+        
+        scoutcommand($directory,$file);
     }
 
 }
@@ -18,14 +18,41 @@ closedir(DIR);
 
 
 sub scoutcommand {
-    my $yaml_file = shift;
+    my ($directory, $file) = @_;
+
+    my $yaml_file = $directory."/".$file;
     my $command = "ssh viktor\@cmdscout1.lund.skane.se 'scout load case $yaml_file'";
     my $log = '/fs1/results/cron/scout/scout_upload.log';
+    my $errlog = '/fs1/results/cron/scout/scout_upload.errlog';
     my $datestring = localtime();
     open(LOG, '>>' , $log) or die $!;
-    print LOG "$datestring :: $yaml_file was loaded using: $command\n\n";
-    my $go = `$command`;
+    
+    
+    my $results = `$command`;
+    my $status = $results->success;
+
+    if ( $status ) {
+        print LOG "$datestring :: $yaml_file was loaded using: $command\n\n";
+        unlink $yaml_file;
+    }
+    else {
+        my $infile = 0;
+        print "HAPPENS\n";
+        open(ERRLOG, $errlog) or die $!;
+        while (<ERRLOG>) {
+            print $file,"\n";
+            if ($_ =~ /\/$file:/) {
+                $infile = 1;
+            }
+        }
+        close(ERRLOG);
+        open(ERRLOG, '>>' , $errlog) or die $!;
+        if ($infile == 0) {
+            print ERRLOG "$datestring :: $yaml_file: could not be loaded.\n";
+        }
+        close(ERRLOG);
+    }
     close(LOG);
-    unlink $yaml_file;
+    
     
 }
